fix subprocess execution (environment)
env={} as parameter defines a mutable default and therefore leads to
funny behaviour of this function. therefore it should be None initialized.
besides that setting env={} is dropping PATH and therefore prevents from
execution of any subprocess that needs PATH resolution.

This change assumes os.environ to be a good default if environment is
unspecified.